github: split helpers
