+++ internal/github/ensure-auth.go
@@ -3,24 +3,7 @@ package github
 import (
 	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
 )
-
-const (
-	authTokenLocation = ".elastic"
-	authTokenFilename = "github.token"
-)
-
-// TokenLocation returns the expected location of the GitHub Token file.
-func TokenLocation() (string, error) {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-
-	return filepath.Join(homeDir, authTokenLocation, authTokenFilename), nil
-}
 
 // EnsureAuthConfigured method ensures that GitHub auth token is available.
 func EnsureAuthConfigured(tk AuthToken) (bool, error) {
